requirements.txt
tensorflow_cpu==2.15.0
matplotlib
numpy
mediapy